tidy: Eliminate some narrowing conversions from double to float.
Mark constants as floating point instead of double.  This keeps the
entire operation in the floating point space, instead of converting
the argument to double to perform the math operation and then
converting the result back to a float.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/narrowing-conversions.html

+++ mythtv/libs/libmyth/audio/freesurround_decoder.cpp
@@ -35,7 +35,7 @@ using OutputBufs = std::array<float*,6>;
 
 static const float PI = 3.141592654;
 static const float epsilon = 0.000001;
-static const float center_level = 0.5*sqrt(0.5);
+static const float center_level = 0.5F*sqrtf(0.5F);
 
 template <class T>
 T sqr(T x) { return x*x; }
